Updated Project struct impl
