+++ src/backend/mod.rs
@@ -8,8 +8,11 @@ pub mod setup;
 pub use error::*;
 
 mod project;
+mod template;
+mod utils;
 
 pub use project::*;
+pub use template::*;
 
 /// Types to operate on
 #[derive(Debug)]
