--- src/backend/project/mod.rs
@@ -1,3 +1,0 @@
-mod project;
-
-pub use project::*;
