+++ src/backend/remove.rs
@@ -2,11 +2,6 @@ use crate::{setup, Type};
 use anyhow::Result;
 use std::fs;
 
-/// Given the name of a project, unregister it.
-pub fn remove_project(name: &String) -> Result<()> {
-    unregister(Type::Project, name)
-}
-
 /// Given the name of a template, unregister it.
 pub fn remove_template(name: &String) -> Result<()> {
     unregister(Type::Template, name)
